‘Amazon Products Purchase Prediction and Comparing ARIMA and RNN Models’
In this project, the main purpose is to predict whether the Purchases for the month is correct, so that which products are purchased more or low in the month of October 2020.  In this project, to work on the simpler problem that is to predict the purchases for the month using two models and finding out the best.